requirements.txt
flask
flask-sqlalchemy
flask-wtf
flask-login
flask-bcrypt
python-dotenv
psycopg2-binary
psycopg2-binary
firebase-admin==6.2.0
python-dotenv==1.0.0
Flask==2.3.2
flask-bcrypt==1.0.1
flask-login==0.6.2
gunicorn==20.1.0
Werkzeug==2.3.7
wtforms==3.0.1
flask-wtf==1.1.1
google-cloud-firestore==2.11.0
reportlab